+++ metametameta/__about__.py
@@ -12,10 +12,10 @@ __all__ = [
 ]
 
 __title__ = "metametameta"
-__version__ = "0.1.4"
+__version__ = "0.1.5"
 __description__ = "Generate __about__.py with dunder meta."
-__credits__ = "[{'name': 'Matthew Martin', 'email': 'matthewdeanmartin@gmail.com'}]"
+__credits__ = [{"name": "Matthew Martin", "email": "matthewdeanmartin@gmail.com"}]
 __readme__ = "README.md"
 __requires_python__ = ">=3.8"
 __keywords__ = ["packaging", "metadata"]
-__status__ = "4 - Beta"
+__status__ = "5 - Production/Stable"