+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.4] - 2024-06-14
+## [0.1.5] - 2024-10-02
+
+### Fixed
+
+- Mypy complaints when lists are empty. Added more null safety checks.
+
+### Added
+
+- `metametameta sync-check` run on CI server to see if your about file is out of sync
+- `metametameta auto` make best guess what your metadata source is. Give up if there are two possibilities.
+
+
+## [0.1.4] - 2024-10-02
 
 ### Fixed
 
@@ -15,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Experimental setup.py support.
 
-## [0.1.3] - 2024-06-14
+## [0.1.3] - 2024-06-28
 
 ### Fixed
 
@@ -27,7 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix bug in pep621 command.
 
-## [0.1.1] - 2024-01-20
+## [0.1.1] - 2024-08-03
 
 ### Added